requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
xformers>=0.0.20
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
flask>=2.3.0
requests>=2.31.0
runpod>=1.5.0
safetensors>=0.3.0
controlnet-aux>=0.0.6
insightface>=0.7.3
onnxruntime>=1.15.0
gradio>=3.40.0
